LabelDetection:translations/requirements.txt
flask
gunicorn
google-cloud-vision==2.6.2
google-cloud-translate==2.0.1
google-cloud-storage==1.42.2
google-cloud-datastore==2.3.0
